Move matcher selection to `MatcherFactory`
After this code move, all matcher selection / creation logic will be
located in `MatcherFactory` only instead of being spread over `Method-
Call`, too.
